requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.35.28
botocore==1.35.99
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.7
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
openai==1.106.1
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==1.26.20

# WhisperX dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.19.0
faster-whisper>=0.9.0
pyannote-audio>=3.1.0

# Clinical Transcription Pipeline dependencies  
librosa>=0.10.0
soundfile>=0.12.0
boto3>=1.35.0
# flash-attn>=2.0.0  # Optional optimization - skip if installation fails
# langdetect removed - language detection now handled during transcription
